# random: Optimize Randomizer::getBytes()